Implementações: Yup para validação de CPF e Email. validação de erro do Yup.
Usuários não podem se excluirem, apenas o proprio usuario pode excluir sua conta.
Se tiver um destino associado a conta, o usuario não pode excluir.

+++ src/routes/user.routes.js
@@ -1,19 +1,27 @@
 const { Router } = require("express");
 const Usuario = require("../models/Usuario");
-const { auth } = require("../middleware/auth") 
-const { verify } = require("jsonwebtoken"); 
+const Destino = require("../models/Destino");
+const { auth } = require("../middleware/auth");
+const { verify } = require("jsonwebtoken");
+const yup = require("yup");
 
 const userRoute = new Router();
 
+const userSchema = yup.object().shape({
+  name: yup.string().required(),
+  gender: yup.string().required(),
+  cpf: yup.string().required().matches(/^\d+$/).length(11),
+  address: yup.string().required(),
+  email: yup.string().email().required(),
+  password: yup.string().required().min(6),
+  birth: yup.date().required(),
+});
+
 userRoute.post("/", async (req, res) => {
   try {
-    const name = req.body.name;
-    const gender = req.body.gender;
-    const cpf = req.body.cpf;
-    const address = req.body.address;
-    const email = req.body.email;
-    const password = req.body.password;
-    const birth = req.body.birth;
+    await userSchema.validate(req.body, { abortEarly: false });
+
+    const { name, gender, cpf, address, email, password, birth } = req.body;
 
     if (!name) {
       return res.status(400).json({ mensagem: "O nome é obrigatório" });
@@ -21,6 +29,15 @@ userRoute.post("/", async (req, res) => {
 
     if (!gender) {
       return res.status(400).json({ mensagem: "Selecione seu genero" });
+    }
+
+    const existingUserCPF = await Usuario.findOne({
+      where: {
+        cpf: cpf,
+      },
+    });
+    if (existingUserCPF) {
+      return res.status(400).json({ message: "CPF já cadastrado" });
     }
 
     if (!cpf) {
@@ -37,6 +54,15 @@ userRoute.post("/", async (req, res) => {
       });
     }
 
+    const existingUserEmail = await Usuario.findOne({
+      where: {
+        email: email,
+      },
+    });
+    if (existingUserEmail) {
+      return res.status(400).json({ message: "Email já cadastrado" });
+    }
+
     const usuario = await Usuario.create({
       name,
       gender,
@@ -48,42 +74,49 @@ userRoute.post("/", async (req, res) => {
     });
     res.status(201).json(usuario);
   } catch (error) {
-    console.log(error.message);
-    res.status(500).json({
-      error: "Não foi possível cadastrar o usuario.",
-    });
+    if (error instanceof yup.ValidationError) {
+      const yupErrors = error.inner.map((e) => ({
+        field: e.path,
+        message: e.message,
+      }));
+      return res.status(400).json({ errors: yupErrors });
+    } else {
+      console.log(error.message);
+      return res.status(500).json({
+        error: "Não foi possível cadastrar o usuário.",
+      });
+    }
   }
 });
 
 userRoute.get("/", async (req, res) => {
   try {
-    
     if (req.headers.authorization) {
-      const token = req.headers.authorization.split(" ")[1]; 
-      const decodedToken = verify(token, process.env.SECRET_JWT); 
-      
-      
+      const token = req.headers.authorization.split(" ")[1];
+      const decodedToken = verify(token, process.env.SECRET_JWT);
+
       if (!decodedToken) {
         return res.status(401).json({ message: "Token inválido" });
       }
-      
-      const userId = decodedToken.sub; 
+
+      const userId = decodedToken.sub;
       const usuario = await Usuario.findByPk(userId);
 
       if (!usuario) {
         return res.status(404).json({ erro: "Usuário não encontrado" });
       }
 
-      return res.json(usuario); 
+      return res.json(usuario);
     } else {
-      
       const usuarios = await Usuario.findAll();
 
-      return res.json(usuarios); 
+      return res.json(usuarios);
     }
   } catch (error) {
     console.log(error.message);
-    return res.status(500).json({ error: "Não foi possível buscar os usuários" });
+    return res
+      .status(500)
+      .json({ error: "Não foi possível buscar os usuários" });
   }
 });
 
@@ -127,15 +160,38 @@ userRoute.put("/:id", auth, async (req, res) => {
 userRoute.delete("/:id", auth, async (req, res) => {
   try {
     const { id } = req.params;
-    const usuario = await Usuario.findByPk(id);
+
+    const userIdFromToken = req.payload.sub;
+    const userIdToDelete = req.params.id;
+
+    if (parseInt(userIdFromToken) !== parseInt(userIdToDelete)) {
+      return res
+        .status(403)
+        .json({ error: "Você não tem permissão para excluir este usuário!" });
+    }
+
+    const usuario = await Usuario.findByPk(userIdToDelete);
 
     if (!usuario) {
       return res.status(404).json({ erro: "Usuário não foi encontrado" });
     }
 
-    Usuario.destroy({
+    const destinosAssociados = await Destino.findAll({
       where: {
-        id: id,
+        user_id: id,
+      },
+    });
+    if (destinosAssociados.length > 0) {
+      return res.status(400).json({
+        error:
+          "Não é possível excluir o usuário, pois existem destinos associados a ele",
+      });
+    }
+
+
+    await Usuario.destroy({
+      where: {
+        id: userIdToDelete,
       },
     });
